refactor(frontend): registry-driven ModalHost; App.tsx 1,131 → 668 lines
All configuration modals now render through modalRegistry + ModalHost
(lazy-loaded); AppHeader and CanvasArea extracted; DeployPanel chat tab
extracted to ChatInterface (964 → 882). Zero behavior change; 240 tests,
tsc, eslint, build all green.
